replace legacy setup.py with pyproject.toml other changes which where required: - move configure_invokeai.py into ldm.invoke - update files which imported configure_invokeai to use new location:     - ldm/invoke/CLI.py     - scripts/load_models.py     - scripts/preload_models.py - update test-invoke-pip.yml:     - remove pr type "converted_to_draft"     - remove reference to dev/diffusers     - remove no more needed requirements from matrix     - add pytorch to matrix     - install via `pip3 install --use-pep517 .`     - use the created executables         - this should also fix configure_invoke not executed in windows To install use `pip install --use-pep517 -e .` where `-e` is optional
